AGS.Native: use common methods to import compiled games into Editor
This makes AGS.Native library use shared MainGameFile functionality instead of its own custom code when importing game data from the binary file.
This essentially lets Editor to import (more or less) any version of AGS game, starting from 2.50, same as Engine.
