+++ Editor/AGS.Native/agsnative.h
@@ -30,7 +30,6 @@
 #define MAXMULTIFILES 25
 #define CHUNKSIZE 256000
 #define MAX_FILENAME_LENGTH 100
-#define SAVEBUFFERSIZE 5120
 #define MAX_PLUGINS 40
 
 extern const char *sprsetname;
